+++ QRhunter/app/build.gradle
@@ -39,6 +39,7 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     implementation platform('com.google.firebase:firebase-bom:31.2.2')
     implementation 'com.google.firebase:firebase-firestore'
+    implementation 'com.google.code.gson:gson:2.9.0'
 
     //Firebase testing
     //testImplementation 'org.powermock:powermock:1.6.5'
