+++ QRhunter/app/src/main/java/com/example/qrhunter/UserPageActivity.java
@@ -1,0 +1,45 @@
+package com.example.qrhunter;
+
+import android.content.Intent;
+import android.util.Log;
+import android.widget.ImageView;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import android.os.Bundle;
+import android.widget.TextView;
+
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
+import com.google.gson.Gson;
+
+import java.util.HashMap;
+
+public class UserPageActivity extends AppCompatActivity {
+
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_user_page);
+
+        // Code for setting the profile circle (can be adapted for custom profile pics later)
+        ImageView profileCircle = (ImageView) findViewById(R.id.default_profile_icon);
+        profileCircle.setImageResource(R.drawable._icon__profile_circle_);
+
+        Gson gson = new Gson();
+        Bundle bundle = getIntent().getExtras();
+        Player p = gson.fromJson(bundle.getString("player", "default"), Player.class);
+        TextView userText = findViewById(R.id.user_page_user_name);
+
+        userText.setText(p.getUsername());
+
+
+
+
+
+
+
+    }
+}
