+++ QRhunter/app/src/main/res/layout/activity_user_page.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".UserPageActivity">
 
 
     <TextView
@@ -17,20 +17,20 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintBottom_toBottomOf="parent" app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="0.242" android:id="@+id/user_name"
+            app:layout_constraintVertical_bias="0.242" android:id="@+id/user_page_user_name"
     />
     <ImageView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content" app:srcCompat="@drawable/_icon__profile_circle_"
             android:id="@+id/default_profile_icon"
             app:layout_constraintStart_toStartOf="parent" app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintBottom_toTopOf="@+id/user_name" app:layout_constraintTop_toTopOf="parent"/>
+            app:layout_constraintBottom_toTopOf="@+id/user_page_user_name" app:layout_constraintTop_toTopOf="parent"/>
     <LinearLayout
             android:orientation="vertical"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             app:layout_constraintEnd_toEndOf="parent" app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/user_name" app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/user_page_user_name" app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintVertical_bias="0.183"
             android:id="@+id/linearLayout">
 
@@ -38,7 +38,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_margin="27dp"
-                android:id="@+id/user_total_score"
+                android:id="@+id/user_page_total_score"
                 android:text="LMAO might want to make this into a fragment">
 
         </TextView>
@@ -46,7 +46,7 @@
         <ListView
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:id="@+id/user_stats_display">
+                android:id="@+id/user_page_stats_display">
 
         </ListView>
     </LinearLayout>
