+++ QRhunter/app/src/test/java/com/example/qrhunter/DatabaseTest.java
@@ -65,7 +65,7 @@ public class DatabaseTest {
     @Test
     public void testGettingPlayerFromUsername(){
         Player p = new Player();
-        p.setUsername("qwertyuiopasdfghjklzxcvbnm"); //Invalid username
+
         db.addPlayer(p)
                 .addOnSuccessListener(new OnSuccessListener<Void>() {
                     @Override
@@ -101,42 +101,61 @@ public class DatabaseTest {
     }
 
     @Test
-    public void testAddScannedCode(){
-        Database db = new Database();
-        Player p = new Player();
-        QRCode qr = new QRCode("name", "location", 123);
-        db.addPlayer(p);
-        db.addQrCode(qr);
-        HashMap<String, Task<Void>> tasks = db.addScannedCode(qr, p);
-        tasks.get("QrToPlayerCol")
-                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void unused) {
-                        Log.d("Db Test", "qr -> player added succesfully");
-                    }
-                })
-                .addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        Log.d("Db Test", "qr -> player Exception: " + e.getMessage());
-                    }
-                });
-        System.out.println(tasks.keySet().toArray()[0]);
-        System.out.println(tasks.keySet().toArray()[1]);
-        tasks.get("PlayerToQrCol")
-                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void unused) {
-                        Log.d("Db Test", "player -> qr added succesfully");
-                    }
-                })
-                .addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        Log.d("Db Test", "player -> qr Exception: " + e.getMessage());
-                    }
-                });
+    public void testAddScannedCode() {
 
+        for (int i = 0; i < 10; i++) {
+
+            int randomScore = (int) Math.floor(Math.random() * 1001); // Generates random score
+            int randomUsername = (int) Math.floor(Math.random() * 1001); // Generates random username (int for now)
+            int randomQRname = (int) Math.floor(Math.random() * 1001); // Generates random QR name (int for now)
+
+            Database db = new Database(); // Creates database
+            Player testPlayer = new Player(String.valueOf(randomUsername)); // Creates player for testing
+            QRCode testQRCode = new QRCode(String.valueOf(randomQRname), "location", randomScore); // Creates QRCode for testing
+            db.addPlayer(testPlayer); // Adds player to database
+            db.addQrCode(testQRCode); // Adds QRCode to database
+
+            /*
+              "tasks" is a hashmap containing two keys, "QrToPlayerCol" and "PlayerToQrCol",
+              which are .set() commands that execute whenever you call their respective keys
+              (hence the success and failure listeners)
+             */
+            HashMap<String, Task<Void>> tasks = db.addScannedCode(testQRCode, testPlayer);
+
+            // Adds a QRCode to the players collection
+            tasks.get("QrToPlayerCol")
+                    .addOnSuccessListener(new OnSuccessListener<Void>() {
+                        @Override
+                        public void onSuccess(Void unused) {
+                            Log.d("Db Test", "qrCode -> playerCollection added successfully");
+                        }
+                    })
+                    .addOnFailureListener(new OnFailureListener() {
+                        @Override
+                        public void onFailure(@NonNull Exception e) {
+                            Log.d("Db Test", "qrCode -> playerCollection Exception: " + e.getMessage());
+                        }
+                    });
+
+            // Log statements for testing
+            Log.d("Db Test", (String) tasks.keySet().toArray()[0]);
+            Log.d("Db Test", (String) tasks.keySet().toArray()[1]);
+
+            // Adds a player to the QRCode collection
+            tasks.get("PlayerToQrCol")
+                    .addOnSuccessListener(new OnSuccessListener<Void>() {
+                        @Override
+                        public void onSuccess(Void unused) {
+                            Log.d("Db Test", "player -> qrCollection added succesfully");
+                        }
+                    })
+                    .addOnFailureListener(new OnFailureListener() {
+                        @Override
+                        public void onFailure(@NonNull Exception e) {
+                            Log.d("Db Test", "player -> qrCollection Exception: " + e.getMessage());
+                        }
+                    });
+        }
     }
 
     /**
